LOG-1441: Creating an operator_sdk bundle release on quay.io, or anywhere.
Described in docs/how-to-deploy-a-bundle.md, in summary:

1. Copy and edit config/overlays/default e.g. as config/overlays/custom
2. make OVERLAY=config/overlays/custom push-bundle

This will build & push the operator and bundle images. The bundle
image is named after the operator with "-bundle" added.

NOTE: Make sure your target repositories are PUBLIC. quay.io repos are
PRIVATE by default and you won't be able to pull from the cluster.